@@ -1,9 +1,22 @@
 module github.com/cosmos/tools/cmd/runsim
 
-go 1.13
+go 1.19
 
 require (
 	github.com/aws/aws-sdk-go v1.23.17
 	github.com/cosmos/tools/lib/runsimgh v1.0.0
 	github.com/cosmos/tools/lib/runsimslack v1.0.0
+)
+
+require (
+	github.com/bradleyfalzon/ghinstallation v0.1.2 // indirect
+	github.com/cosmos/tools/lib/runsimaws v1.0.0 // indirect
+	github.com/dgrijalva/jwt-go v3.2.0+incompatible // indirect
+	github.com/google/go-github/v27 v27.0.6 // indirect
+	github.com/google/go-querystring v1.0.0 // indirect
+	github.com/gorilla/websocket v1.2.0 // indirect
+	github.com/jmespath/go-jmespath v0.0.0-20180206201540-c2b33e8439af // indirect
+	github.com/nlopes/slack v0.6.0 // indirect
+	github.com/pkg/errors v0.8.0 // indirect
+	golang.org/x/crypto v0.0.0-20190308221718-c2843e01d9a2 // indirect
 )
